move bigquery test to unit tests
